Changes with nginx 1.5.5                                         17 Sep 2013

*) Change: now nginx assumes HTTP/1.0 by default if it is not able to
detect protocol reliably.

*) Feature: the "disable_symlinks" directive now uses O_PATH on Linux.

*) Feature: now nginx uses EPOLLRDHUP events to detect premature
connection close by clients if the "epoll" method is used.

*) Bugfix: in the "valid_referers" directive if the "server_names"
parameter was used.

*) Bugfix: the $request_time variable did not work in nginx/Windows.

*) Bugfix: in the "image_filter" directive.
Thanks to Lanshun Zhou.

*) Bugfix: OpenSSL 1.0.1f compatibility.
Thanks to Piotr Sikora.

Conflicts:
	src/core/nginx.h

+++ src/http/modules/ngx_http_proxy_module.c
@@ -1732,7 +1732,7 @@ ngx_http_proxy_chunked_filter(ngx_event_pipe_t *p, ngx_buf_t *buf)
 
             if (buf->last - buf->pos >= ctx->chunked.size) {
 
-                buf->pos += ctx->chunked.size;
+                buf->pos += (size_t) ctx->chunked.size;
                 b->last = buf->pos;
                 ctx->chunked.size = 0;
 
@@ -1895,7 +1895,7 @@ ngx_http_proxy_non_buffered_chunked_filter(void *data, ssize_t bytes)
             b->tag = u->output.tag;
 
             if (buf->last - buf->pos >= ctx->chunked.size) {
-                buf->pos += ctx->chunked.size;
+                buf->pos += (size_t) ctx->chunked.size;
                 b->last = buf->pos;
                 ctx->chunked.size = 0;
